extract ValidateAndHandleCommand & add exception logging
